pyproject.toml
[project]
name = "music-catalog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
]
